move normalizr, initialized data to cv-lib/constants
