+++ vue/package.json
@@ -13,7 +13,6 @@
   "dependencies": {
     "chart.js": "^2.7.1",
     "github-colors": "^2.2.16",
-    "normalizr": "^3.2.4",
     "vue": "^2.5.2",
     "vue-chartjs": "^3.1.1",
     "vue-moment": "^3.1.0",
